Former-commit-id: 138aeb36d225a2b94625f81188db166c23a0d3f9

+++ backend/src/main/java/com/moi/anitime/api/request/bookmark/BookmarkReq.java
@@ -13,7 +13,7 @@ public class BookmarkReq {
     private int generalNo;
 
     @NotEmpty
-    private int desertionNo;
+    private long desertionNo;
 
     public Bookmark toEntity() {
         Bookmark bookmark = Bookmark.builder()
